Add clipboard utility with fallback and show selectable URLs
